LINUX: Zero code on EEXIST in afs_linux_read_cache
In afs_linux_read_cache(), if we get a -EEXIST error from
afs_add_to_page_cache_lru(), we'll effectively ignore the error, and go
through the surrounding while() loop again to retry getting the relevant
page. But when we do so, 'code' is still set to -EEXIST. So if
find_get_page() now returns non-NULL, we'll break out of the while()
loop, and continue through the rest of the function with 'code' still
set. We won't update the given page, and we'll never set 'code' to
anything else, eventually returning the confusing error -EEXIST.

While triggering this may be a hard race to hit (it would require the
backing page to suddenly exist between find_get_page() and
afs_add_to_page_cache_lru()) and hasn't been known to actually happen,
this seemingly could cause a afs_linux_readpage()/afs_linux_read_folio()
to fail unexpectedly, possibly causing a spurious SIGBUS, which would be
hard to identify.

To avoid this, set 'code' to 0 after examining 'code' after the
afs_add_to_page_cache_lru() call. Add some braces around the nearby 'if'
to help make it clear this is outside of that block.

This issue appears to have existed since this function was introduced in
commit f75d84c404 (Add support for blocking readahead).

Change-Id: Ib1a3ee0dfaa29c9cb7a087e6cc822a81206ef89f
Reviewed-on: https://gerrit.openafs.org/16590
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>