auth: Introduce afsconf_PickClientLocalSecObj()
Introduce new function afsconf_PickClientLocalSecObj(), which is the
same as afsconf_PickClientSecObj(), except that it is only used for
localauth. This new function lacks the parameters for the cell (info,
cellName), but gains a parameter for the identity to become (user). For
now, this function only supports RX_ID_KRB4 and RX_ID_SUPERUSER users.

To implement this, move most code from afsconf_PickClientSecObj() to a
new internal function PickClientSecObj(). This new function is the same
as afsconf_PickClientSecObj(), but uses GenericAuth() instead of
afsconf_ClientAuth/afsconf_ClientAuthSecure, so we can provide the given
rx_identity.

Change-Id: Ic9ca56f27bf03fa92c9e3bb8bf123b57a55cc490
Reviewed-on: https://gerrit.openafs.org/14570
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>