libafscp: Support rxkad_krb5 keys for local auth
Add support for rxkad_krb5 keys for local auth, allowing authentication
as arbitrary users using local key material compatible with rxkad_krb5.

To achieve this, _GetLocalSecurityObject() is updated to delegate the
authentication of the connection to afsconf_PickLocalSecObj(), which is
rxkad_krb5 friendly.

The afscp_LocalAuthAs() function is also updated to reflect that we now
pass a single string to _GetLocalSecurityObject(), and the authas_name
buffer size is increased to 512 bytes to accommodate both the name and
the instance.

Change-Id: I82663f1647be34e7554e52edb1ddcf9840281e10
Reviewed-on: https://gerrit.openafs.org/14568
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>