tests: Use plain $? in perl tests
Some perl tests currently use ${^CHILD_ERROR_NATIVE} with
WEXITSTATUS/WTERMSIG to check if a child process died with a certain
exit status or signal. This works, but adds a few lines of unnecessary
complexity; using plain $? and encoding the relevant values is much more
concise. Using ${^CHILD_ERROR_NATIVE} also requires perl 5.10; if using
an older perl (such as 5.008), this variable expands to nothing, which
confusingly results in WEXITSTATUS() et al thinking at the process
exited with exit status 0. This currently causes the rx/simple-t test to
fail with perl 5.008 (such as on AIX 6), because we expect the server
process to be killed with SIGTERM.

Some other tests already use $?, but throw away some information (e.g.,
we look at '$? & 0x7f' or '$? >> 8'). If the process actually exited
when we expected it to die from a signal, or died from a signal when we
expected it to exit, the TAP output won't show what actually happened.

To avoid all of this, change perl tests to test against the plain $?
value as much as possible, so the TAP output will just show the encoded
$? value if the child process has an unexpected status. We still clear
the WCOREDUMP bit for some users, since we don't care whether its set,
but otherwise we test against the plain $? value. This isn't as pretty,
but it concisely shows almost all information about the exit status,
which is very helpful when something goes wrong.

Change-Id: I68fd140aab7e65bec1ac537d19de89f9d32443c1
Reviewed-on: https://gerrit.openafs.org/16594
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>