tests: Use pthread-based binaries instead of LWP
The test suite currently runs the LWP versions of vos and pt_util
instead of the pthreaded ones. On macOS arm64, the LWP implementation
has been unreliable for various reasons, often causing segmentation
faults during execution.

While those are bugs and could be fixed, LWP in general is deprecated,
and we care more about the pthreaded versions. At least for now, just
update these tests to run the pthreaded versions of binaries, so we can
run the tests on more platforms, and so we test the binaries that are
more likely to actually be used.

Change-Id: I27038a075511a6bc1379bae72a51995b6c222a13
Reviewed-on: https://gerrit.openafs.org/16596
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>