pts: Free pr_NameToId(), pr_IdToName() outputs
The GetNameOrId() function calls pr_NameToId() and pr_IdToName() to
gather the name and id lists, but does not free the output arguments
from these functions, so the xdr allocated results are leaked.

Be sure to free the output arguments here to avoid memory leaks each
time the -nameorid switch is processed in pts.

Change-Id: I0081581a254f4a417202070455c20062bc56b95f
Reviewed-on: https://gerrit.openafs.org/16597
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>