pts: Free intermediate lists in GetNameOrId()
The GetNameOrId() function allocates a pair of lists called `ids` and
`names` to gather the names and ids from the ptserver for each name or
id given in the -nameorid command line option.  The final results are
copied to the output arguments `lids` and `lnames`, but the internal
`ids` and `name` lists are not freed.  Free these internal lists to
avoid leaking memory every time GetNameOrId() is called.

Also, free the `lnames` output list when an error is encountered. The
`lids` is already freed on errors, but it appears `lnames` was missed.

This change fixes a memory leak in pts for every command which has the
-nameorid switch.

For example, when running `pts membership` with valgrind:

    4 bytes in 1 blocks are definitely lost in loss record 6 of 547
       at 0x4846828: malloc (in /usr/libexec/valgrind/vgpreload_memcheck-amd64-linux.so)
       by 0x11034B: GetNameOrId (in /usr/local/src/openafs/src/tptserver/pts)
       by 0x1109FD: ListMembership (in /usr/local/src/openafs/src/tptserver/pts)
       by 0x17265E: cmd_Dispatch (in /usr/local/src/openafs/src/tptserver/pts)
       by 0x112AA4: main (in /usr/local/src/openafs/src/tptserver/pts)

    64 bytes in 1 blocks are definitely lost in loss record 499 of 547
       at 0x484D953: calloc (in /usr/libexec/valgrind/vgpreload_memcheck-amd64-linux.so)
       by 0x110367: GetNameOrId (in /usr/local/src/openafs/src/tptserver/pts)
       by 0x1109FD: ListMembership (in /usr/local/src/openafs/src/tptserver/pts)
       by 0x17265E: cmd_Dispatch (in /usr/local/src/openafs/src/tptserver/pts)
       by 0x112AA4: main (in /usr/local/src/openafs/src/tptserver/pts)

Change-Id: I99aaadb753021b982b72802797f9a0b944e50e54
Reviewed-on: https://gerrit.openafs.org/16598
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>