afsio: Exit on CmdProlog error
CmdProlog() can return an error, but none of its callers check what it
returns. Change all callers to check for an error from CmdProlog(), and
return an error if so. Return -1 on error, which is what various code
paths in here already do for internal errors.

CmdProlog() never actually returns an error right now (unless the given
cmd_syndesc is NULL, which cannot happen), but future commits may add
code paths that return an error.

Change-Id: Ia10aadc9e73219c68a82e91872484091a42b3beb
Reviewed-on: https://gerrit.openafs.org/16600
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>