afs: Add new pioctl PUnlogCell
The current implementation of PUnlog removes all tokens and it does not
provide a mechanism to remove tokens for a specific cell. In a multi-cell
environment, this behavior is inefficient because it removes tokens for
all cells, even when only a single cell’s tokens need to be invalidated.

To address this limitation, introduce a new pioctl, PUnlogCell
(pioctl code O.3), which removes tokens only for the specified cell.
This avoids unintended token removal for other cells and eliminates the
need for additional pioctl calls to re-acquire tokens that are
unnecessarily revoked by PUnlog.

Change-Id: I23ee8bde66e1e743febe22ae7e4ec822ff2ae7e0
Co-authored-by: Koustubh Mattikalli <koustubh.mattikalli@ibm.com>
Co-authored-by: Amar Bhandari <amar.bhandari@ibm.com>
Reviewed-on: https://gerrit.openafs.org/16620
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>