INSTALL: Update AIX section to clarify compiler support
Update the INSTALL document to clarify the compiler support
situation.  We currently support the original IBM XL C
family, and at this time, Open XL C 17.1.0 and 17.1.1 on
AIX 7.2+.  As IBM still uses XL C for the system compiler
on these AIX releases, indicate that XL C is the preferred
solution at this time in order to not need to update this
document on a regular basis.

Change-Id: I5621a74e402f85668aca59f1ed453eb588a852ad
Reviewed-on: https://gerrit.openafs.org/16636
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>