pts: Remove remnant help command check in CleanUp()
The CleanUp() function in pts.c has a remnant check to return early when
passed a cmd syntax descriptor named "help".  This is dead code left
over from various changes over the the years and can be removed at this
point.

The GetGlobals() already returns early if the command is "help", so the
call to CleanUp() in GetGlobals() never hits that code path.  The calls
to CleanUp() in main() always pass NULL arguments, so the CleanUp()
never checks for a "help" command in those code paths.

Originally, CleanUp() was a libcmd "after proc" hook. The commit
8ab7a90937 (pts-supergroups-20030114) added the Source and Interactive
commands, we required the hook to be removed, since pts was changed to
call cmd_Dispatch() multiple times after the initialization performed in
the before hook (GetGlobals).  The original hooks had a check to avoid
initialization and shutdown when the command was "help". This check is
still present in the GetGlobals() "before hook", but is no longer needed
in CleanUp(), since it is no longer a hook.

Remove the leftover check for "help" and change the CleanUp() signature,
since it is called directly now, instead of from the "after proc" hook.

Change-Id: Idf53123ddc1f7dd0d50fc03b46e327d8cc9f5178
Reviewed-on: https://gerrit.openafs.org/16641
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>