pts: Add recursion depth limit to the source command
A 'source' command can create an infinite recursive loop. This can
happen directly, where a file sources itself, or indirectly, where a
chain of source commands eventually calls back to one of the files in
the chain (e.g., a.txt -> b.txt -> a.txt).

When this happens, pts will be terminated by the operating system when
the number of open files is exceeded or the process runs out or memory,
leading to a confusing error message.

Instead, limit the recursion depth of the 'source' command to prevent
this resource exhaustion. If the limit is exceeded, make 'source' return
an error and print an error message:

    Error sourcing 'filename': Source command recursion depth limit (64) exceeded.

Add a statement to the pts_source man page to let users know about
the recursion limit.

Change-Id: I8e74864a18a75e8d31373b5be4ed7ea861b6cab3
Reviewed-on: https://gerrit.openafs.org/16643
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>