build: Remove `which` dependency for configure generation
The build scripts used to generate the `configure` script currently
depend on the `which` command.  In minimal Linux environments (such as
containers), the `which` command may not be installed.

Instead of running `which`, update regen.sh to check for the libtoolize
and aclocal commands by running them with --version to check if they are
available.  Update the git-version command to run git with the 'git
version' command to check if it is available.

This allows us to run regen.sh on a minimal system that has git,
autoconf, and libtool install, but is missing the `which` command.

Change-Id: Idfc9de46e77be3767512bae6a0a2e8b6b7d8d9bf
Reviewed-on: https://gerrit.openafs.org/16645
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>