cf: Add /usr/src/kernels to linux headers search path
Add the modern '/usr/src/kernels/$(uname -r)' to the list of paths to
search for the Linux headers.  This is the path used by RedHat based
distros since RHEL 7.

The /lib/modules/$(uname -r)/build path we currently rely on by default
is usually a symlink to the modern '/usr/src/kernels/$(uname -r)' path,
but the /lib/modules path may not be present, for example, when building
in a container.

This change improves the detection of the Linux headers path when the
user did not specify a path with configure options.

Change-Id: I7de0edac219284536485b73bd3988f9df35b2143
Reviewed-on: https://gerrit.openafs.org/16649
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>