cf: Fix check for ENABLE_FUSE_CLIENT
Commit 2e5545342b (fuse: Autodetect Solaris 11 FUSE) checked if we
failed to detect fuse by checking for

    [test x"ENABLE_FUSE_CLIENT" = x]

Which is always false, instead of:

    [test x"$ENABLE_FUSE_CLIENT" = x]

As a result, if someone explicitly enabled the fuse client by passing
--enable-fuse-client to configure, and fuse headers are not installed,
the build would just continue without fuse support.

Fix this by adding the missing '$', and so now 'configure' fails if fuse
cannot be found and --enable-fuse-client was given.

Change-Id: Ib55ce70d3f910a8ab1448edb3e4b4bc91fe239e9
Reviewed-on: https://gerrit.openafs.org/16665
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Gaurav Saxena <gsaxena@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>