cf: Report fuse errors with $FUSE_PKG_ERRORS
Commit 2e5545342b (fuse: Autodetect Solaris 11 FUSE) tries to report
errors detecting fuse by printing $FUSE_PACKAGE_ERRORS, but our pkg.m4
stores errors in variable $FUSE_PKG_ERRORS. As a result, if you run
configure with --enable-fuse-client and fuse is not installed, you get a
rather confusing error:

    $ ./configure --enable-fuse-client
    [...]
    checking for FUSE... no
    configure: error: ""

To fix this, print $FUSE_PKG_ERRORS instead, and remove the unnecessary
extra quotes. This gives us a better error message:

    $ ./configure --enable-fuse-client
    [...]
    checking for FUSE... no
    configure: error: Package 'fuse', required by 'virtual:world', not found

This is still maybe not the clearest error for unfamiliar users, but
this is what pkg.m4 gives us, and is certainly at least an improvement.

Change-Id: I3058cffab805145a8d9b7b078a4becf6508fa3ca
Reviewed-on: https://gerrit.openafs.org/16666
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Gaurav Saxena <gsaxena@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>